Moved password hashing functions to Utils
Moved password hashing functions to Utils
