+++ app/src/main/java/com/myapps/toualbiamine/food2class/SignIn.java
@@ -18,6 +18,7 @@ import com.google.firebase.auth.*;
 import com.google.firebase.database.*;
 import com.myapps.toualbiamine.food2class.Common.Common;
 import com.myapps.toualbiamine.food2class.Model.User;
+import com.myapps.toualbiamine.food2class.Utils.PasswordHasher;
 import com.rengwuxian.materialedittext.MaterialEditText;
 
 import org.libsodium.jni.NaCl;
@@ -49,6 +50,8 @@ public class SignIn extends AppCompatActivity {
     ImageButton closePopupBtn;
     EditText forgotPwdEmail;
 
+    PasswordHasher passwordHasher;
+
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -61,6 +64,7 @@ public class SignIn extends AppCompatActivity {
         btnSignIn = (Button) findViewById(R.id.signInBtn);
         rememberMeCb = (CheckBox) findViewById(R.id.rememberMeCb);
         forgotPassword = (TextView) findViewById(R.id.forgotPassword);
+        passwordHasher = new PasswordHasher();
 
         signInProgressBar = (ProgressBar) findViewById(R.id.signInProgressBar);
         signInProgressBar.setVisibility(View.INVISIBLE);
@@ -178,8 +182,8 @@ public class SignIn extends AppCompatActivity {
 
 
                         //if(!(user.getPassword().equals(signInPassword))) {
-                        if(!checkPassword(signInPassword,user.getPassword())){
-                            String encryptedPassword = hashPassword(signInPassword);
+                        if(!passwordHasher.checkPassword(signInPassword,user.getPassword())){
+                            String encryptedPassword = passwordHasher.hashPassword(signInPassword);
                             user.setPassword(encryptedPassword);
                             tableUser.child(signInEmail).setValue(user);
                         }
@@ -293,44 +297,5 @@ public class SignIn extends AppCompatActivity {
         return formatted;
 
     }
-    //checks password and returns false if the password is incorrect
-    public boolean checkPassword(String guess,String theHashedPassword){
-        byte[] passBytes = guess.getBytes();
-        byte[] hashedPass = stringToByteArray(theHashedPassword);
-        if(NaCl.sodium().crypto_pwhash_scryptsalsa208sha256_str_verify(hashedPass,passBytes,passBytes.length)!=0){
-            //wrong password
-            return false;
-        }
-        //right password
-        return true;
-    }
-    private String hashPassword(String password){
-        byte[] hashedPass = new byte[NaCl.sodium().crypto_pwhash_strbytes()];
-        byte[] passBytes = password.getBytes();
-        NaCl.sodium().crypto_pwhash_scryptsalsa208sha256_str(hashedPass,passBytes,passBytes.length,NaCl.sodium().crypto_pwhash_scryptsalsa208sha256_opslimit_interactive(),NaCl.sodium().crypto_pwhash_scryptsalsa208sha256_memlimit_interactive());
-        return byteArrayToString(hashedPass);
-    }
-    //converts a byte array into a special string that can be undone by stringToByteArray
-    private String byteArrayToString(byte[] a){
-        String s = "";
-        for(byte c:a){
-            s+=c+",";
-        }
-        return s;
-    }
-    //converts a string made from byteArrayToString back into a byte array
-    private byte[] stringToByteArray(String encryptedPassword){
-        byte[] hashedPass = new byte[NaCl.sodium().crypto_pwhash_strbytes()];
-        String[] splitString = encryptedPassword.split(",");
-        if(splitString.length == hashedPass.length){
-            for(int i=0;i<splitString.length;i++) {
-                hashedPass[i]=Byte.parseByte(splitString[i]);
-            }
 
-        }
-        else{
-            System.out.println("Why aren't they the same size?!");
-        }
-        return hashedPass;
-    }
 }
