+++ app/src/main/java/com/myapps/toualbiamine/food2class/SignUp.java
@@ -17,6 +17,7 @@ import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
 import com.google.firebase.database.*;
 import com.myapps.toualbiamine.food2class.Model.User;
+import com.myapps.toualbiamine.food2class.Utils.PasswordHasher;
 import com.rengwuxian.materialedittext.MaterialEditText;
 
 import org.libsodium.jni.NaCl;
@@ -42,6 +43,8 @@ public class SignUp extends AppCompatActivity {
     String signUpPassword;
     String signUpName;
 
+    PasswordHasher passwordhasher;
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -50,6 +53,7 @@ public class SignUp extends AppCompatActivity {
         emailInput = (MaterialEditText) findViewById(R.id.emailSignUp);
         nameInput = (MaterialEditText) findViewById(R.id.nameSignUp);
         passwordInput = (MaterialEditText) findViewById(R.id.passwordSignUp);
+        passwordhasher = new PasswordHasher();
 
         signUpBtn = (Button) findViewById(R.id.signUpBtn);
 
@@ -174,7 +178,7 @@ public class SignUp extends AppCompatActivity {
 
                 else {
 
-                    String encryptedPassword = hashPassword(signUpPassword);
+                    String encryptedPassword = passwordhasher.hashPassword(signUpPassword);
 
                     User newUser = new User(signUpEmail, signUpName, encryptedPassword);
                     tableUser.child(signUpEmail).setValue(newUser);
@@ -276,36 +280,5 @@ public class SignUp extends AppCompatActivity {
 
         return formatted.toLowerCase();
 
-    }
-
-    private String hashPassword(String password){
-
-        byte[] hashedPass = new byte[NaCl.sodium().crypto_pwhash_strbytes()];
-        byte[] passBytes = password.getBytes();
-        NaCl.sodium().crypto_pwhash_scryptsalsa208sha256_str(hashedPass,passBytes,passBytes.length,NaCl.sodium().crypto_pwhash_scryptsalsa208sha256_opslimit_interactive(),NaCl.sodium().crypto_pwhash_scryptsalsa208sha256_memlimit_interactive());
-        return byteArrayToString(hashedPass);
-    }
-    //converts a byte array into a special string that can be undone by stringToByteArray
-    private String byteArrayToString(byte[] a){
-        String s = "";
-        for(byte c:a){
-            s+=c+",";
-        }
-        return s;
-    }
-    //converts a string made from byteArrayToString back into a byte array
-    private byte[] stringToByteArray(String encryptedPassword){
-        byte[] hashedPass = new byte[NaCl.sodium().crypto_pwhash_strbytes()];
-        String[] splitString = encryptedPassword.split(",");
-        if(splitString.length == hashedPass.length){
-            for(int i=0;i<splitString.length;i++) {
-                hashedPass[i]=Byte.parseByte(splitString[i]);
-            }
-
-        }
-        else{
-            System.out.println("Why aren't they the same size?!");
-        }
-        return hashedPass;
     }
 }
